project/requirements.txt
gunicorn==20.0.4
celery==4.4.6
Django==3.0.6
django-celery-beat==2.0.0
django-js-asset==1.2.2
django-timezone-field==4.0
redis==3.5.3
requests==2.23.0
schedule==0.6.0
djangorestframework==3.11.0
redis==3.5.3
psycopg2-binary==2.8.5
Pillow==7.1.2
beautifulsoup4==4.9.1
fake_useragent==0.1.11